Move logic to asset-controller
