+++ components/asset-controller/.gitignore
