+++ components/asset-controller/Dockerfile
