+++ components/asset-controller/config/crds/objectstore_v1alpha1_bucket.yaml
