+++ components/asset-controller/config/default/kustomization.yaml
