+++ components/asset-controller/config/default/manager_image_patch.yaml
