+++ components/asset-controller/config/default/manager_prometheus_metrics_patch.yaml
