+++ components/asset-controller/config/manager/manager.yaml
