+++ components/asset-controller/config/rbac/auth_proxy_role.yaml
