+++ components/asset-controller/config/rbac/rbac_role.yaml
