+++ components/asset-controller/config/rbac/rbac_role_binding.yaml
