+++ components/asset-controller/config/samples/objectstore_v1alpha1_bucket.yaml
