+++ components/asset-controller/pkg/apis/objectstore/v1alpha1/register.go
