+++ components/asset-controller/pkg/buckethandler/automock/minio_client.go
