+++ components/asset-controller/pkg/controller/add_bucket.go
@@ -1,7 +1,7 @@
 package controller
 
 import (
-	"github.com/kyma-project/kyma/components/bucket-controller/pkg/controller/bucket"
+	"github.com/kyma-project/kyma/components/asset-controller/pkg/controller/bucket"
 )
 
 func init() {
